Add files via upload
Hey there, I am Sadia Kauser. This is a simple chatbot that will save your to-do-list in a sql database . Just get your Bot-Token from telegram and edit it in main code , help.py is a class which will help creating the table. Try it , and comments section is always open for suggestions.